Refactor Text module
